requirements.txt
streamlit
transformers
torch
google-api-python-client
python-dotenv
wordcloud
plotly
matplotlib
pandas
